requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.20.0

# Computer vision and image processing
pillow>=9.5.0
opencv-python>=4.8.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Machine learning utilities
datasets>=2.14.0
tokenizers>=0.13.0

# Model-specific dependencies
qwen-vl-utils>=0.0.8

# Evaluation and metrics
scikit-learn>=1.3.0

# Development and utilities
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and environment
python-dotenv>=1.0.0
omegaconf>=2.3.0

# Optional: For specific models (uncomment if needed)
# flash-attn>=2.0.0  # For attention optimization
# xformers>=0.0.20   # For memory efficient attention

# Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0 